Refactor GameArea into a container and a component
