+++ test/reactHelpers.js
@@ -23,7 +23,7 @@ export function renderElement(elt, deep = false) {
 /* eslint-disable react/no-multi-comp */
 
 export function createGameArea(state, dispatch = noop, props = {}) {
-  return React.createElement(gameArea.GameArea, {...gameArea.mapStateToProps(state), ...gameArea.mapDispatchToProps(dispatch), ...props});
+  return React.createElement(gameArea.GameAreaContainer, {...gameArea.mapStateToProps(state), ...gameArea.mapDispatchToProps(dispatch), ...props});
 }
 
 function createCreator(state, dispatch = noop) {
